requirements.txt
matplotlib
numpy
tensorflow
opencv-python
Pillow
torch
torchvision
albumentations
